Updating Matlab Time Embedding
The Dynamics Matrix Regression function has been updated to appropriately stack the kronecker products vertically rather than horizontally.
This results in a regression to A and B vectorized matrices that are at least non-zero and more in line with the process that I believe Zac originally described.

The Time Delay Embed function has been updated to output a true Hankel matrix in that the first columns are not couched with a constant initial condition if the time delay pre-dates the passed in matrix index. The quantity of its output columns are updated based on the quantity of time delays requested.

A Time Embed Forward Rollout function has been added to calculate a simulated set of jaw pressures given a trajectory of inputs, initial condition, and dynamic system matrices. This can be used to check simulated results of the modeled A and B matrices versus actual experimental data. Uncertain if this is yet fully working as its results with time delay embedded systems are unstable which is counter-intuitive to how they should behave.

The Hankel Exploration main function has been updated to use these helper functions accordingly and additionally has had some plotting code starting to be added to aid in visualizing results.