Updates to move away from vec/kron
Hankel Exploration is still the parent / umbrella file. It uses Dynamics_Mat_Regression with a call that doesn't use time delay embedding on the control input, only the state.

Additionally, the Dynamics Matrix Regression function has been updated to linearly solve for a B matrix and A_i matrices that correspond to how many time delays are in place.

A 'no_kron_forward_rollout' function was created to account for the different structure of the system now for rolling it out.